Test HFNoseTriggerDetId code

+++ DataFormats/ForwardDetId/src/HFNoseTriggerDetId.cc
@@ -55,7 +55,7 @@ int HFNoseTriggerDetId::triggerCellX() const {
   for (auto const& v : vc) {
     x += (3 * (v - N) + 2);
   }
-  return (x / vc.size());
+  return (x / static_cast<int>(vc.size()));
 }
 
 int HFNoseTriggerDetId::triggerCellY() const {
@@ -67,7 +67,7 @@ int HFNoseTriggerDetId::triggerCellY() const {
   for (unsigned int k = 0; k < uc.size(); ++k) {
     y += (2 * uc[k] - (N + vc[k]));
   }
-  return (y / vc.size());
+  return (y / static_cast<int>(vc.size()));
 }
 
 std::vector<int> HFNoseTriggerDetId::cellU() const {
